api/pyproject.toml
[project]
name = "35gateway-console-api"
version = "0.1.0"
description = "35gateway console backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.135.2",
  "uvicorn[standard]>=0.42.0",
  "sqlalchemy>=2.0.48",
  "pydantic-settings>=2.13.1",
  "httpx>=0.28.1",
  "aiosqlite>=0.21.0",
  "scalar-fastapi>=1.8.1",
  "python-multipart>=0.0.22",
  "jinja2>=3.1.6",
  "wechatpayv3>=2.0.2",
  "tencentcloud-sdk-python-sms>=3.0.1400",
  "tencentcloud-sdk-python-ses>=3.0.1400",
  "dashscope>=1.25.15",
  "greenlet>=3.3.2",
  "openai>=2.29.0",
  "ddgs>=9.11.4",
  "nest-asyncio>=1.6.0",
  "playwright>=1.58.0",
  "oss2>=2.19.1",
  "PyPDF2>=3.0.1",
  "Pillow>=12.1.1",
  "beautifulsoup4>=4.14.3",
  "html2docx>=1.6.0",
  "google-genai>=1.68.0",
  "httpx-socks>=0.11.0",
  "shadowsocks @ git+https://github.com/shadowsocks/shadowsocks.git@5ff694b2c2978b432918dea6ac104706b25cbf48",
]

[project.scripts]
console-api = "app.main:run"
console-api-stable = "app.main:run_stable"

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
  "ruff>=0.15.7",
  "mypy>=1.19.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "ipython>=9.10.0",
]
